cache: move methods used for testing into test file
